feat: Add Welcome component and update CSS configurations
This commit includes the addition of a new Welcome component and the modification of the tailwind configuration to include a new color. Styles for the Welcome component have been created in a new CSS module. In src/App.jsx, the interface has been updated to conditionally render the Welcome component based on the state.